Made FeatureExtractor a separate file added documentation to featureExtractor.py to help people customize it
